requirements.txt
asgiref==3.6.0
Flask==3.0.3
matplotlib==3.9.2
numpy==1.23.5
opencv_python==4.8.0.74
Pillow==10.4.0
scikit_learn==1.4.1.post1
xgboost==2.1.1
